Move the auction loop out of lib.rs into a run_loop module
